dev: abobus

+++ modifiers/scale.go
@@ -11,11 +11,11 @@ import (
 type ScaleX struct {
 	// TODO: embed?
 	In    *webp.Animation
-	Scale float32
+	Scale float64
 }
 
 func (m ScaleX) Modify() (*webp.AnimationEncoder, error) {
-	newWidth := int(float32(m.In.CanvasWidth) * m.Scale)
+	newWidth := int(float64(m.In.CanvasWidth) * m.Scale)
 	stride := newWidth * 4
 
 	enc, err := webp.NewAnimationEncoder(newWidth, m.In.CanvasHeight, 0, 0)
@@ -28,7 +28,7 @@ func (m ScaleX) Modify() (*webp.AnimationEncoder, error) {
 		for j := 0; j < m.In.CanvasHeight; j++ {
 			for i := 0; i < newWidth; i++ {
 				for k := 0; k < 4; k++ {
-					buf[j*stride+i*4+k] = frame.Pix[j*frame.Stride+int(float32(i)/m.Scale)*4+k]
+					buf[j*stride+i*4+k] = frame.Pix[j*frame.Stride+int(float64(i)/m.Scale)*4+k]
 				}
 			}
 		}
@@ -50,11 +50,11 @@ func (m ScaleX) Modify() (*webp.AnimationEncoder, error) {
 
 type ScaleY struct {
 	In    *webp.Animation
-	Scale float32
+	Scale float64
 }
 
 func (m ScaleY) Modify() (*webp.AnimationEncoder, error) {
-	newHeight := int(float32(m.In.CanvasHeight) * m.Scale)
+	newHeight := int(float64(m.In.CanvasHeight) * m.Scale)
 
 	enc, err := webp.NewAnimationEncoder(m.In.CanvasWidth, newHeight, 0, 0)
 	if err != nil {
@@ -66,7 +66,7 @@ func (m ScaleY) Modify() (*webp.AnimationEncoder, error) {
 		for j := 0; j < newHeight; j++ {
 			for i := 0; i < m.In.CanvasWidth; i++ {
 				for k := 0; k < 4; k++ {
-					buf[j*frame.Stride+i*4+k] = frame.Pix[int(float32(j)/m.Scale)*frame.Stride+i*4+k]
+					buf[j*frame.Stride+i*4+k] = frame.Pix[int(float64(j)/m.Scale)*frame.Stride+i*4+k]
 				}
 			}
 		}
@@ -88,7 +88,7 @@ func (m ScaleY) Modify() (*webp.AnimationEncoder, error) {
 
 type ScaleT struct {
 	In    *webp.Animation
-	Scale float32
+	Scale float64
 }
 
 func (m ScaleT) Modify() (_ *webp.AnimationEncoder, e error) {
@@ -103,7 +103,7 @@ func (m ScaleT) Modify() (_ *webp.AnimationEncoder, e error) {
 	}()
 
 	for i, frame := range m.In.Image {
-		if err := enc.AddFrame(frame, time.Duration(float32(m.In.Timestamp[i])*m.Scale)*time.Millisecond); err != nil {
+		if err := enc.AddFrame(frame, time.Duration(float64(m.In.Timestamp[i])*m.Scale)*time.Millisecond); err != nil {
 			return nil, err
 		}
 	}
